renamed instance list tables to be more appropriate w/ burnettk
