+++ spiffworkflow-frontend/src/routes/Extension.tsx
@@ -26,7 +26,7 @@ import ProcessInstanceRun from '../components/ProcessInstanceRun';
 import MarkdownRenderer from '../components/MarkdownRenderer';
 import LoginHandler from '../components/LoginHandler';
 import SpiffTabs from '../components/SpiffTabs';
-import ProcessInstanceListTable from '../components/ProcessInstanceListTable';
+import ProcessInstanceListTableWithFilters from '../components/ProcessInstanceListTableWithFilters';
 import CreateNewInstance from './CreateNewInstance';
 
 type OwnProps = {
@@ -69,7 +69,7 @@ export default function Extension({
   const supportedComponents: ObjectWithStringKeysAndFunctionValues = {
     CreateNewInstance,
     SpiffTabs,
-    ProcessInstanceListTable,
+    ProcessInstanceListTableWithFilters,
     ProcessInstanceRun,
     MarkdownRenderer,
   };
